+++ src/helpers/index.ts
@@ -1,4 +1,3 @@
-export * from './canActivate';
 export * from './defineMiddleware';
 export * from './isFatcherError';
 export * from './readStreamByChunk';
