+++ src/index.ts
@@ -1,5 +1,4 @@
+export * from './core';
 export * from './helpers';
-export { FatcherError } from './core/FatcherError';
-export { combine } from './utils';
 export * from './interfaces';
-export * from './fatcher';
+export * from './middlewares';
