+++ src/middlewares/json/json.ts
@@ -1,4 +1,6 @@
-import { canActivate, defineMiddleware, Middleware } from '../..';
+import { defineMiddleware } from '../../helpers';
+import { Middleware } from '../../interfaces';
+import { canActivate } from '../../core';
 
 /**
  * A middleware for transforming stream into json
